fix: restore review regression audit import
